library_backend/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.1
django-ninja==1.3.0
dnspython==2.7.0
email_validator==2.2.0
gevent==24.11.1
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
ninja-put-patch-file-upload-middleware==0.1.4
orjson==3.10.15
packaging==24.2
pillow==11.1.0
psycopg2==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
python-decouple==3.8
requests==2.32.3
setuptools==75.8.0
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
whitenoise==6.9.0
zope.event==5.0
zope.interface==7.2
